requirements.txt
akshare>=1.12.90
efinance>=0.5.5
pandas>=2.0.0
PyYAML>=6.0
python-dotenv>=1.0.0
colorlog>=6.7.0
requests>=2.31.0
google-generativeai>=0.3.0
tenacity>=8.2.0

pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
mypy>=1.0.0
